refactor(data_operations): group arithmetic shared code into arithmetic/ package
The arithmetic families' shared code (the ArithmeticFeatureGroupBase
skeleton, the five per-backend mixins, and the five per-backend
numeric-source descriptors) sat flat at the data_operations root, mixed
in with cross-cutting helpers. All of it is arithmetic-only, so move it
into a self-contained arithmetic/ subpackage:

  arithmetic_base.py            -> arithmetic/base.py
  {backend}_arithmetic_mixin.py -> arithmetic/{backend}_mixin.py
  {backend}_numeric_source.py   -> arithmetic/{backend}_numeric_source.py

Pure relocation: class names, constants, and logic are unchanged; only
module paths move. All importers (the two family bases, the ten concrete
backends, the testing bases, and the tests' static and dynamic import
paths) are updated accordingly.
