+++ mloda/community/feature_groups/data_operations/arithmetic/__init__.py
@@ -1,0 +1,1 @@
+"""Shared arithmetic building blocks (base, per-framework mixins, numeric-source descriptors)."""
