+++ mloda/community/feature_groups/data_operations/arithmetic/duckdb_numeric_source.py
