+++ mloda/community/feature_groups/data_operations/arithmetic/pandas_mixin.py
@@ -15,8 +15,10 @@ import pandas as pd
 from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.pandas.dataframe import PandasDataFrame
 
-from mloda.community.feature_groups.data_operations.arithmetic_base import ArithmeticFeatureGroupBase
-from mloda.community.feature_groups.data_operations.pandas_numeric_source import pandas_non_numeric_descriptor
+from mloda.community.feature_groups.data_operations.arithmetic.base import ArithmeticFeatureGroupBase
+from mloda.community.feature_groups.data_operations.arithmetic.pandas_numeric_source import (
+    pandas_non_numeric_descriptor,
+)
 
 
 class PandasArithmeticMixin(ArithmeticFeatureGroupBase):
