+++ mloda/community/feature_groups/data_operations/arithmetic/sqlite_numeric_source.py
